Updated 1)dependencies 2)int to long in entities 3)constants for content database
Added
1)DAOs
2)Result class with Success and Error

Fixed bug with crashed application when not found map file for maps fragment

+++ app/src/main/java/com/grsu/guideapp/database/content/entities/Lines.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 class Lines(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id_line")
-    val id_line: Int,
+    val id_line: Long,
 
     @ColumnInfo(name = "start_point") val start_point: String,
     @ColumnInfo(name = "end_point") val end_point: String,
