+++ app/src/main/java/com/grsu/guideapp/database/content/entities/ListPoi.kt
@@ -25,8 +25,8 @@ import androidx.room.ForeignKey
     ]
 )
 class ListPoi(
-    @ColumnInfo(name = "id_route", index = true) val id_route: Int,
-    @ColumnInfo(name = "id_poi", index = true) val id_poi: Int,
+    @ColumnInfo(name = "id_route", index = true) val id_route: Long,
+    @ColumnInfo(name = "id_poi", index = true) val id_poi: Long,
     @ColumnInfo(name = "id_point") val id_point: String,
     @ColumnInfo(name = "distance") val distance: Int
 )
