+++ app/src/main/java/com/grsu/guideapp/database/content/entities/ReferencesList.kt
@@ -25,6 +25,6 @@ import androidx.room.ForeignKey
     ]
 )
 class ReferencesList(
-    @ColumnInfo(name = "id_poi") val id_poi: Int,
+    @ColumnInfo(name = "id_poi") val id_poi: Long,
     @ColumnInfo(name = "id_reference") val id_reference: String
 )
